requirements.txt
fastapi
uvicorn
prisma
python-jose
bcrypt
passlib
python-dotenv